+++ plugins/module_utils/ksconf_shared.py
@@ -7,9 +7,7 @@ __metaclass__ = type
 import os
 import re
 
-from ansible.errors import AnsibleActionFail
 from ansible.module_utils.basic import AnsibleModule
-from ansible.utils.display import Display
 
 
 __version__ = "0.19.0"
@@ -25,16 +23,43 @@ SPLUNK_HOME_PATH = [
 
 
 def find_splunk_home():
+    """ Find an appropriate value for SPLUNK_HOME.
+
+    First, use $SPLUNK_HOME if that directory exists.  After that, try a series
+    of path-based guesses based on common installation files, and finally just
+    go with *any* match against the well known splunk home list.
+
+    Multiple splunk installations are not explicitly supported.
+    """
     if "SPLUNK_HOME" in os.environ:
-        return os.environ["SPLUNK_HOME"]
-    for path in SPLUNK_HOME_PATH:
+        path = os.environ["SPLUNK_HOME"]
         if os.path.isdir(path):
             return path
+
+    # Check popular paths
+    for known_path in ("bin/splunk", "etc/splunk.version", None):
+        path = _guess_splunk_home(SPLUNK_HOME_PATH, known_path)
+        if path:
+            return path
+    return None
+
+
+def _guess_splunk_home(discovery_paths, test_file):
+    for path in discovery_paths:
+        if os.path.isdir(path):
+            if test_file:
+                test_path = os.path.join(path, test_file)
+                if os.path.isfile(test_path):
+                    return path
+            else:
+                return path
     return None
 
 
 def check_ksconf_version(module: AnsibleModule = None) -> tuple:
     if not module:
+        from ansible.errors import AnsibleActionFail
+        from ansible.utils.display import Display
         display = Display()
     try:
         from ksconf import __version__ as ksconf_version